debugging data_processing and branch tests solved

+++ src/assembler/define_types.h
@@ -78,7 +78,8 @@ typedef struct {
 typedef struct {
     bool immediate;
     union {
-        uint8_t expression;
+        //uint8_t expression;
+        int expression;
         struct {
             uint8_t rm;
             Shift shift;
